test(mcp): isolate press key validation suite
